requirements.txt
aniso8601==9.0.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
munch==2.5.0
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pymodm==0.4.3
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.2.1
zipp==3.8.1
dnspython==2.2.1